Author: hno Fix some minor casting errors affecting cachemgr reporting when cache/mem >2GB

+++ src/stat.cc
@@ -1,5 +1,5 @@
 /*
- * $Id: stat.cc,v 1.410 2007/09/20 20:22:20 hno Exp $
+ * $Id: stat.cc,v 1.410.2.1 2008/02/25 02:42:43 amosjeffries Exp $
  *
  * DEBUG: section 18    Cache Manager Statistics
  * AUTHOR: Harvest Derived
@@ -541,16 +541,16 @@ info_get(StoreEntry * sentry)
                       store_swap_size);
 
     storeAppendPrintf(sentry, "\tStorage Swap capacity:\t%4.1f%% used, %4.1f%% free\n",
-                      dpercent((int) store_swap_size, (int) Store::Root().maxSize()),  
-                      dpercent((int) (Store::Root().maxSize() - store_swap_size), (int) Store::Root().maxSize()));
+                      dpercent(store_swap_size, Store::Root().maxSize()),  
+                      dpercent((Store::Root().maxSize() - store_swap_size), Store::Root().maxSize()));
 
 
     storeAppendPrintf(sentry, "\tStorage Mem size:\t%d KB\n",
-                      (int) (mem_node::store_mem_size >> 10));
+                      (mem_node::store_mem_size >> 10));
 
     storeAppendPrintf(sentry, "\tStorage Mem capacity:\t%4.1f%% used, %4.1f%% free\n",
                       dpercent(mem_node::InUseCount(), store_pages_max),
-                      dpercent(((double)store_pages_max - mem_node::InUseCount()), store_pages_max));
+                      dpercent((store_pages_max - mem_node::InUseCount()), store_pages_max));
 
     storeAppendPrintf(sentry, "\tMean Object Size:\t%0.2f KB\n",
                       n_disk_objects ? (double) store_swap_size / n_disk_objects : 0.0);
